requirements.txt
beautifulsoup4==4.9.2
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
convertdate==2.2.2
cycler==0.10.0
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
holidays==0.10.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.2
pandas==0.25.3
pandas-datareader==0.9.0
plotly==4.10.0
PyMeeus==0.3.7
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
tqdm==4.49.0
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
